Remove custom DataSet class.
