+++ src/continuity/data/datasets.py
@@ -3,10 +3,73 @@
 import torch
 import numpy as np
 from typing import Tuple
-from continuity.data import DataSet, tensor
+from continuity.data import tensor
+from torch import Tensor
+from torch.utils.data import Dataset
 
 
-class Sine(DataSet):
+class SelfSupervisedDataSet(Dataset):
+    """
+    A `SelfSupervisedDataSet` is a data set that exports observations and labels
+    for self-supervised learning. Every data point is created by taking one
+    sensor as label.
+
+    Every sample consists of tuples `(x, u, y, v)`, where `x` contains the sensor
+    positions, `u` the sensor values, and `y = x_i` and `v = u_i` are
+    the label's coordinate its value for all `i`.
+
+    Args:
+        x: Sensor positions of shape (num_observations, num_sensors, coordinate_dim)
+        u: Sensor values of shape (num_observations, num_sensors, num_channels)
+    """
+
+    def __init__(
+        self,
+        x: Tensor,
+        u: Tensor,
+    ):
+        self.num_observations = u.shape[0]
+        self.num_sensors = u.shape[1]
+        self.coordinate_dim = x.shape[-1]
+        self.num_channels = u.shape[-1]
+
+        # Check consistency across observations
+        for i in range(self.num_observations):
+            assert (
+                x[i].shape[-1] == self.coordinate_dim
+            ), "Inconsistent coordinate dimension."
+            assert (
+                u[i].shape[-1] == self.num_channels
+            ), "Inconsistent number of channels."
+
+        xs, us, ys, vs = [], [], [], []
+
+        for i in range(self.num_observations):
+            # Add one data point for every sensor
+            for j in range(self.num_sensors):
+                y = x[i][j].unsqueeze(0)
+                v = u[i][j].unsqueeze(0)
+
+                xs.append(x[i])
+                us.append(u[i])
+                ys.append(y)
+                vs.append(v)
+
+        self.xs = torch.stack(xs)
+        self.us = torch.stack(us)
+        self.ys = torch.stack(ys)
+        self.vs = torch.stack(vs)
+
+        super().__init__()
+
+    def __len__(self) -> int:
+        return len(self.xs)
+
+    def __getitem__(self, i: int) -> Tuple[Tensor, Tensor, Tensor, Tensor]:
+        return self.xs[i], self.us[i], self.ys[i], self.vs[i]
+
+
+class Sine(Dataset):
     r"""Creates a data set of sine waves.
 
     The data set is generated by sampling sine waves at the given number of
@@ -36,9 +99,7 @@ class Sine(DataSet):
         shuffle: Shuffle data set. Defaults to True.
     """
 
-    def __init__(
-        self, num_sensors: int, size: int, batch_size: int = 32, shuffle: bool = True
-    ):
+    def __init__(self, num_sensors: int, size: int):
         self.num_sensors = num_sensors
         self.size = size
 
@@ -48,14 +109,20 @@ class Sine(DataSet):
         # Generate observations
         observations = [self.generate_observation(i) for i in range(self.size)]
 
-        x = torch.stack([x for x, _ in observations])
-        u = torch.stack([u for _, u in observations])
+        self.x = torch.stack([x for x, _ in observations])
+        self.u = torch.stack([u for _, u in observations])
 
         # Use observations as labels
-        y = x
-        v = u
+        self.y = self.x
+        self.v = self.u
 
-        super().__init__(x, u, y, v, batch_size, shuffle)
+        super().__init__()
+
+    def __len__(self) -> int:
+        return self.size
+
+    def __getitem__(self, i: int) -> Tuple[Tensor, Tensor, Tensor, Tensor]:
+        return self.x[i], self.u[i], self.y[i], self.v[i]
 
     def generate_observation(self, i: float) -> Tuple[np.array, np.array]:
         """Generate observation
